fix(front-end): move file and fix key in menubar
