+++ frontend_web/src/route.js
@@ -1,10 +1,10 @@
-import Home from "../pages/Home/home";
-import DetailUser from "../pages/Account/Detail";
-import NotFound from "../pages/NotFound/notfound";
-import DataTable from "../components/Table/dataTable";
-import CreateUser from "../pages/User/CreateUser";
+import Home from "./pages/Home/home";
+import DetailUser from "./pages/Account/Detail";
+import NotFound from "./pages/NotFound/notfound";
+import DataTable from "./components/Table/dataTable";
+import CreateUser from "./pages/User/CreateUser";
 import { Route, Routes as ReactRoutes } from "react-router-dom";
-import { MenuList } from "../layouts/Sidebar/sidebar";
+import { MenuList } from "./layouts/Sidebar/sidebar";
 
 
 export const getRouterByPath = (path) => {
